+++ react/ImageList.tsx
@@ -2,11 +2,11 @@ import React from 'react'
 import { useDevice } from 'vtex.device-detector'
 import { ListContextProvider, useListContext } from 'vtex.list-context'
 
-import { IMAGE_LIST_SCHEMA } from './utils/schema'
-import { getImagesAsJSXList } from './utils/imageUtils'
+import { ImagesSchema, IMAGE_LIST_SCHEMA } from './modules/schema'
+import { getImagesAsJSXList } from './modules/imageAsList'
 
 interface Props {
-  images: Image[]
+  images: ImagesSchema
   height: number
 }
