+++ react/ImageListWithoutContext.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { useDevice } from 'vtex.device-detector'
 
-import { IMAGE_LIST_SCHEMA } from './utils/schema'
-import { getImagesAsJSXList } from './utils/imageUtils'
+import { ImagesSchema, IMAGE_LIST_SCHEMA } from './modules/schema'
+import { getImagesAsJSXList } from './modules/imageAsList'
 
 interface ImageListProps {
-  images: Image[]
+  images: ImagesSchema
   height: number
 }
